+++ test/test-directory.c
@@ -109,7 +109,7 @@ print_list (GnomeVFSDirectoryList *list)
 			GNOME_VFS_FILE_INFO_SYMLINK (info) ? " [link]" : "",
 			type_to_string (info->type),
 			gnome_vfs_file_info_get_mime_type (info),
-			info->size, info->permissions);
+			(glong) info->size, info->permissions);
 
 		info = gnome_vfs_directory_list_next (list);
 	}
